requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
gunicorn>=22.0.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
pyyaml>=6.0.1

# LLM
litellm>=1.60.0
langfuse>=2.59.7,<3.0.0  # optional analytics — only used when ANALYTICS__ENABLED=true; litellm requires 2.x

# PII
presidio-analyzer>=2.2.354
presidio-anonymizer>=2.2.354
spacy>=3.7.4

# RAG
chromadb>=0.5.0
sentence-transformers>=2.7.0
tree-sitter-languages>=1.10.0  # pre-compiled grammars for 40+ languages

# Database
sqlalchemy[asyncio]>=2.0.30
alembic>=1.13.0
asyncpg>=0.29.0
aiosqlite>=0.20.0

# Metrics
prometheus-client>=0.20.0

# Rate limiting (optional Redis backend)
redis[hiredis]>=5.0.0

# Utilities
python-multipart>=0.0.9
httpx>=0.27.0
tenacity>=8.3.0
structlog>=24.1.0
cachetools>=5.3.0

